+++ apps/decodex/src/mcp/tools.rs
@@ -3,9 +3,9 @@ use serde_json::Value;
 use super::{
 	McpCapabilityProfile, McpTool, TOOL_AUTONOMY_ACCEPT_OBJECTIVE,
 	TOOL_AUTONOMY_CHALLENGE_PROPOSAL, TOOL_AUTONOMY_COMPILE_PROPOSAL,
-	TOOL_AUTONOMY_DRAFT_OBJECTIVE, TOOL_AUTONOMY_REQUEST_PROMOTION,
-	TOOL_AUTONOMY_SUBMIT_SIGNAL, TOOL_INTAKE_GOAL, TOOL_LANE_CONTROL, TOOL_OBSERVE,
-	TOOL_PLAN, TOOL_PROJECT_CONTROL, TOOL_RESEARCH_COMPILE, TOOL_RESEARCH_PROMOTE,
+	TOOL_AUTONOMY_DRAFT_OBJECTIVE, TOOL_AUTONOMY_REQUEST_PROMOTION, TOOL_AUTONOMY_SUBMIT_SIGNAL,
+	TOOL_INTAKE_GOAL, TOOL_LANE_CONTROL, TOOL_OBSERVE, TOOL_PLAN, TOOL_PROJECT_CONTROL,
+	TOOL_RESEARCH_COMPILE, TOOL_RESEARCH_PROMOTE,
 };
 
 pub(super) fn mcp_tools() -> Vec<McpTool> {
